+++ src/lib/pages/account-details/components/tables/instantiated-contracts/Lite.tsx
@@ -1,8 +1,9 @@
 import { Box } from "@chakra-ui/react";
 
+import { AccountDetailsEmptyState } from "../../AccountDetailsEmptyState";
 import AccountSectionWrapper from "../../AccountSectionWrapper";
 import { useInternalNavigate, useMobile } from "lib/app-provider";
-import { EmptyState, ErrorFetching } from "lib/components/state";
+import { ErrorFetching } from "lib/components/state";
 import { ContractsTable, MobileTitle, ViewMore } from "lib/components/table";
 import { useAccountContractsLcd } from "lib/pages/account-details/data";
 import type { BechAddr32 } from "lib/types";
@@ -44,12 +45,14 @@ export const InstantiatedContractsTableLite = ({
             isLoading={isLoading}
             emptyState={
               !contracts ? (
-                <ErrorFetching dataName="contracts" />
-              ) : (
-                <EmptyState
-                  message="No contracts have been instantiated by this account before."
+                <ErrorFetching
+                  dataName="instantiated contracts"
                   withBorder
+                  my={2}
+                  hasBorderTop={false}
                 />
+              ) : (
+                <AccountDetailsEmptyState message="No contracts have been instantiated by this account before." />
               )
             }
             onRowSelect={onRowSelect}