+++ src/lib/pages/instantiate/InstantiateFormPage.tsx
@@ -56,13 +56,13 @@ import type { BechAddr, BechAddr20, BechAddr32, ComposedMsg } from "lib/types";
 import { MsgType } from "lib/types";
 import {
   composeMsg,
-  isId,
   jsonPrettify,
   jsonValidate,
   libDecode,
   resolvePermission,
 } from "lib/utils";
 
+import { zInstantiateQueryParams } from "./types";
 import type { InstantiateFormState, InstantiateRedoMsg } from "./types";
 
 interface InstantiateFormPageProps {
@@ -80,8 +80,9 @@ const InstantiateFormPage = ({ onComplete }: InstantiateFormPageProps) => {
   // ---------------DEPENDENCIES---------------//
   // ------------------------------------------//
   const router = useRouter();
-  const msgQuery = (router.query.msg as string) ?? "";
-  const codeIdQuery = (router.query["code-id"] as string) ?? "";
+  const { msg: msgQuery, codeId: codeIdQuery } = zInstantiateQueryParams.parse(
+    router.query
+  );
   const { user: exampleUserAddress } = useExampleAddresses();
   const { address } = useCurrentChain();
   const postInstantiateTx = useInstantiateTx();
@@ -113,7 +114,7 @@ const InstantiateFormPage = ({ onComplete }: InstantiateFormPageProps) => {
   } = useForm<InstantiateFormState>({
     mode: "all",
     defaultValues: {
-      codeId: "",
+      codeId: 0,
       codeHash: "",
       label: "",
       adminAddress: "",
@@ -144,7 +145,7 @@ const InstantiateFormPage = ({ onComplete }: InstantiateFormPageProps) => {
   // -------------------DATA-------------------//
   // ------------------------------------------//
   const { data: derivedWasmVerifyInfo } = useDerivedWasmVerifyInfo(
-    codeId.length ? Number(codeId) : undefined,
+    codeId,
     codeHash
   );
 
@@ -159,7 +160,7 @@ const InstantiateFormPage = ({ onComplete }: InstantiateFormPageProps) => {
     const generalChecks =
       Boolean(address) &&
       Boolean(label) &&
-      isId(codeId) &&
+      codeId &&
       status.state === "success";
 
     switch (tab) {
@@ -303,7 +304,7 @@ const InstantiateFormPage = ({ onComplete }: InstantiateFormPageProps) => {
       const decodedMsg = libDecode(msgQuery);
       try {
         const msgObject = JSON.parse(decodedMsg) as InstantiateRedoMsg;
-        setValue("codeId", msgObject.code_id.toString());
+        setValue("codeId", msgObject.code_id);
         setValue("label", msgObject.label);
         setValue("adminAddress", msgObject.admin);
         setValue(
@@ -332,10 +333,10 @@ const InstantiateFormPage = ({ onComplete }: InstantiateFormPageProps) => {
   useEffect(() => {
     setValue("codeHash", "");
     setTab(MessageTabs.JSON_INPUT);
-    if (codeId.length) {
+    if (codeId) {
       setStatus({ state: "loading" });
       const timer = setTimeout(() => {
-        if (isId(codeId)) refetch();
+        if (codeId) refetch();
         else setStatus({ state: "error", message: "Invalid Code ID" });
       }, 500);
       return () => clearTimeout(timer);
@@ -353,7 +354,7 @@ const InstantiateFormPage = ({ onComplete }: InstantiateFormPageProps) => {
             composeMsg(MsgType.INSTANTIATE, {
               sender: address,
               admin: adminAddress as BechAddr,
-              codeId: Long.fromString(codeId),
+              codeId: Long.fromInt(codeId),
               label,
               msg: Buffer.from(currentInput),
               funds,
@@ -424,18 +425,18 @@ const InstantiateFormPage = ({ onComplete }: InstantiateFormPageProps) => {
             </Heading>
           )}
           <CodeSelectSection
+            codeId={codeId}
             name="codeId"
             control={control}
-            status={status}
             error={formErrors.codeId?.message}
-            onCodeSelect={(code: string) => {
+            onCodeSelect={(code: number) => {
               setValue("codeId", code);
               resetMsgInputSchema();
             }}
             setCodeHash={(data: Code) =>
               setValue("codeHash", data.hash.toLowerCase())
             }
-            codeId={codeId}
+            status={status}
           />
         </Box>
         <form style={{ width: "100%" }}>
@@ -496,7 +497,7 @@ const InstantiateFormPage = ({ onComplete }: InstantiateFormPageProps) => {
               />
             }
             schemaContent={
-              isId(codeId) && (
+              codeId && (
                 <SchemaInputSection
                   type="instantiate"
                   codeHash={codeHash}