+++ src/lib/pages/txs/index.tsx
@@ -2,7 +2,7 @@ import { useRouter } from "next/router";
 import { useEffect } from "react";
 
 import { AmpEvent, track } from "lib/amplitude";
-import { useTierConfig } from "lib/app-provider";
+import { useEvmConfig, useTierConfig } from "lib/app-provider";
 import PageContainer from "lib/components/PageContainer";
 import { PageHeader } from "lib/components/PageHeader";
 import { CelatoneSeo } from "lib/components/Seo";
@@ -12,9 +12,9 @@ import { TxsTableFull } from "./components/TxsTableFull";
 import { TxsTableSequencer } from "./components/TxsTableSequencer";
 
 const Txs = () => {
-  useTierConfig({ minTier: "sequencer" });
-
   const router = useRouter();
+  useTierConfig({ minTier: "sequencer" });
+  const evm = useEvmConfig({ shouldRedirect: false });
 
   useEffect(() => {
     if (router.isReady) track(AmpEvent.TO_TXS);
@@ -25,8 +25,7 @@ const Txs = () => {
       <CelatoneSeo pageName="Transactions" />
       <PageHeader
         title="Transactions"
-        subtitle="This page displays all transactions on this network sorted by
-        recency"
+        subtitle={`This page displays all ${evm.enabled ? "cosmos" : ""} transactions on this network sorted by recency`}
         docHref="introduction/overview#recent-transactions"
       />
       <TierSwitcher