+++ src/lib/types/evm.ts
@@ -1,0 +1,13 @@
+export enum EvmMethodId {
+  Transfer = "0x",
+  TransferErc20 = "0xa9059cbb",
+  Create = "0x60806040",
+  CallErc20Factory = "0x06ef1a86",
+}
+
+export enum EvmMethodName {
+  Transfer = "Transfer",
+  TransferErc20 = "Transfer ERC20",
+  Create = "Create",
+  CallErc20Factory = "Call ERC20 Factory",
+}