+++ src/lib/types/json.ts
@@ -1,1 +1,4 @@
-export type JsonDataType = unknown;
+import { z } from "zod";
+
+export const zJsonDataType = z.unknown();
+export type JsonDataType = z.infer<typeof zJsonDataType>;