+++ src/lib/utils/formatter/tokenType.ts
@@ -7,6 +7,7 @@ export const getTokenType = (type: string) => {
   switch (type.toLowerCase()) {
     case "ibc":
     case "cw20":
+    case "evm":
       return type.toUpperCase();
     default:
       return capitalize(type);